+++ src/ask_for_help_bot/main.py
@@ -46,6 +46,8 @@ if __name__ == "__main__":
     max_users = int(os.getenv("MAX_USERS", 5))
     survey_url = os.getenv("SURVEY_URL")
     helper_url = os.getenv("PILOT_HELPER_URL")
+    channel_id = os.getenv("CHANNEL_ID")
+    publication_language = os.getenv("PUBLICATION_LANGUAGE", "en")
     wenet_authentication_url = os.getenv("WENET_AUTHENTICATION_URL")
     redirect_url = os.getenv("REDIRECT_URL")
     client_secret = os.getenv("CLIENT_SECRET")
@@ -76,6 +78,8 @@ if __name__ == "__main__":
         max_users,
         survey_url,
         helper_url,
+        channel_id,
+        publication_language,
         alert_module,
         connector,
         None,